# book-network-api/src/main/java/com/springboot/booknetwork/auth/AuthenticationResponse.java
package com.springboot.booknetwork.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {

    private String token;
}