# book-network-api/src/main/java/com/springboot/booknetwork/feedback/FeedbackResponse.java
package com.springboot.booknetwork.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}